refactor: file nameing 형식 변경
- model -> class 따라
- controller -> 소문자
