+++ src/controllers/github.ts
@@ -3,7 +3,7 @@ import { Request, Response } from "express";
 import fetch from "node-fetch";
 import { Model } from "sequelize-typescript";
 import HttpError from "../errors/HttpError";
-import User from "../models/user";
+import User from "../models/User";
 import { generateToken } from "../util/generateToken";
 import { serialize } from "../util/serialize";
