+++ src/router/yearcategory.ts
@@ -1,5 +1,5 @@
 import express, { Router } from "express";
-import { readYearCategory } from "../controllers/YearCategory";
+import { readYearCategory } from "../controllers/yearCategory";
 import wrapAsync from "../util/wrapAsync";
 
 const yearcategoryRouter: Router = express.Router();
